[SPARK-56062][PYTHON] Isolate memory_profiler to improve import time
### What changes were proposed in this pull request?

* Moved `memory_profiler` related code to another file
* Make that module lazy so importing it won't trigger `import memory_profiler`
* Updated the code that uses `memory_profiler` related code to use the new module

### Why are the changes needed?

We should make `import pyspark` fast. Bootstrap time is important to user experience, and our CI speed. Also, we should make worker spawn faster so simple worker can be faster to boot. We have a pending SPIP about using simpler worker for UDF work and import time is critical.

`import pyspark` takes **220 ms** on master and will be **90 ms** after this PR. Similar acceleration will be achieved for `pyspark.worker`.

This would probably make a lot of our tests faster as a side effect. Including the newly added python data source simple worker test which timeout a lot in coverage run - https://github.com/apache/spark/actions/runs/23046212656/job/66936106272 .

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

I ran some profiler tests locally and they passed. CI is needed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #54895 from gaogaotiantian/isolate-memory-profiler.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
